feat(xpath): Support wildcards
