+++ tests/samples/James-LG_Skyscraper.html
@@ -585,7 +585,6 @@
   <details-dialog class="command-palette-details-dialog d-flex flex-column flex-justify-center height-fit" role="dialog" aria-modal="true">
     <command-palette class="command-palette color-bg-default rounded-3" data-return-to="/James-LG/Skyscraper" data-user-id="1709432" data-catalyst="">
       <input type="hidden" value="YRi8uxUy+viMmtLDH8LzsnXxPwaSkczG8pHgeTYV3jzSqV1kdCHCxO9mBVJq2TAoXEKVqvepINyU/Y/m/UXZWQ==" data-csrf="true" class="js-graphql-csrf">
-      <iframe class="d-none js-command-palette-commands" sandbox="allow-scripts allow-same-origin" aria-hidden="true" data-url="https://github.com" src="./James-LG_Skyscraper_files/saved_resource.html"></iframe>
       <input type="hidden" name="color-mode-path" id="color-mode-path" value="/settings/appearance/color_mode" class="js-color-mode-path" autocomplete="off">
       <input type="hidden" value="qXrIhTmLjvU3s5uWZQTQz7oUAVvaCKMtIHMOzlaQcIpR6drtQ2O5Dzq+4wQwySjiIHDePZBbxHeUhPdHjx8hIA==" data-csrf="true" class="js-color-mode-csrf">
